Code Clean up
	- format file structure
	- remove reduntant code code
	- move float to int
