+++ .gitignore
@@ -3,3 +3,4 @@ packages
 bin
 obj
 Properties
+*.pfx
