+++ DiscOut/Avolonia/DiscWindow.axaml
@@ -1,0 +1,19 @@
+﻿<Window xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:controls="clr-namespace:DiscOut.Avalonia"
+        Width="640" Height="720"
+        x:Class="DiscOut.Avalonia.DiscWindow"
+        Title="DISCout" Background="Black">
+  
+  <Border Background="Gray" BorderThickness="20">
+    <Grid>
+      <Panel Background="Black">
+        <controls:GameOpenGLControl Name="GLView" HorizontalAlignment="Stretch" VerticalAlignment="Stretch"/>
+      </Panel>
+      <TextBlock Foreground ="White" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="10" Name="ScoreText"/>
+      <TextBlock Foreground ="White" HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="36" Name="DisplayText"/>
+    </Grid>
+  </Border>
+</Window>
