+++ DiscOut/Avolonia/GameOpenGLControl.axaml
@@ -1,0 +1,4 @@
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             x:Class="DiscOut.Avalonia.GameOpenGLControl">
+</UserControl>
