+++ SquareSmash.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.33808.371
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SquareSmash", "SquareSmash.csproj", "{92802523-BB23-4993-BFD2-8D6513482DA8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{92802523-BB23-4993-BFD2-8D6513482DA8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{92802523-BB23-4993-BFD2-8D6513482DA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{92802523-BB23-4993-BFD2-8D6513482DA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{92802523-BB23-4993-BFD2-8D6513482DA8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {776A66BB-161C-487A-9704-407C5C31DD20}
+	EndGlobalSection
+EndGlobal
