+++ assets/levels/level_2.json
@@ -1,39 +1,39 @@
 ﻿{
-	"BaseBallSpeed": 5,
+	"BaseBallSpeed": 3,
 	"Bricks": [
 		{
 			"Colour": "DiscGreen",
-			"Type": 1,
+			"Type": 0,
 			"Repeat": 28
 		},
 		{
 			"Colour": "DiscOrange",
-			"Type": 1,
+			"Type": 0,
 			"Repeat": 14
 		},
 		{
 			"Colour": "DiscOrange",
-			"Type": 1,
+			"Type": 0,
 			"Repeat": 6
 		},
 		{
 			"Colour": "DiscOrange",
-			"Type": 2,
+			"Type": 0,
 			"Repeat": 2
 		},
 		{
 			"Colour": "DiscOrange",
-			"Type": 1,
+			"Type": 0,
 			"Repeat": 6
 		},
 		{
 			"Colour": "DiscBlue",
-			"Type": 1,
+			"Type": 0,
 			"Repeat": 28
 		},
 		{
 			"Colour": "DiscPink",
-			"Type": 1,
+			"Type": 0,
 			"Repeat": 28
 		}
 	]
