--- assets/shaders/basic_vertex.glsl
@@ -1,20 +1,0 @@
-#version 330 core
-// input
-layout(location = 0) in vec2 position;
-layout(location = 1) in float colour;
-layout (location = 2) in float uv;
-// output
-out vec3 vertex_colour;
-out vec2 texture_coord;
-
-//colour look up
-vec3 colours[] = vec3[](vec3(255,255,255),vec3(157, 6, 241),vec3(255, 127, 80),vec3(251, 0, 250),vec3(0, 192, 237),vec3(249, 185, 0),vec3(0, 238, 0));
-
-vec2 uvs[] = vec2[](vec2(0.0, 0.0),vec2(1.0, 0.0),vec2(0.0, 1.0),vec2(1.0, 1.0));
-
-void main()
-{
-    gl_Position = vec4(position, 0.0, 1.0);
-    vertex_colour = normalize(colours[int(colour)]);
-    texture_coord = uvs[int(uv)];
-}
