--- objects/components/bricks/BrickType.cs
@@ -1,9 +1,0 @@
-﻿namespace SquareSmash.objects.components.bricks
-{
-    public enum BrickType
-    {
-        AIR,
-        NORMAL,
-        LIFE
-    }
-}
