--- renderer/Windows/Controls/GameOpenGLControl.axaml
@@ -1,4 +1,0 @@
-﻿<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             x:Class="SquareSmash.renderer.Windows.Controls.GameOpenGLControl">
-</UserControl>
